Изменены значения рейтинга

+++ Model/Toilet.cs
@@ -6,7 +6,7 @@
         public string? Name;
         public Location? Location;
         public string? Description;
-        public short Raiting;
+        public double Raiting;
 
         public bool IsValid()
         {
